requirements.txt
pytest
pandas
simple_salesforce
mysql-connector-python
